Structural refactor: slim facade, partitioner-owned binning, one-shot removal

+++ earthcatalog/config.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 @dataclass
 class GridConfig:
     type: str = "h3"
-    resolution: int | None = None  # H3 / S2 resolution level
+    resolution: float | None = None  # H3/S2 level; lat_lon: degrees per tile
     boundaries_path: str | None = None  # GeoJSON partitioner: path to boundaries file
     id_field: str | None = None  # GeoJSON partitioner: property to use as key
     time_bin: str = "year"  # "year" | "month" | "day" — temporal binning
